Created .xcodeproj to enable Carthage support
