+++ SwiftyUserDefaults.podspec
@@ -1,6 +1,6 @@
 Pod::Spec.new do |s|
   s.name = 'SwiftyUserDefaults'
-  s.version = '1.1.0'
+  s.version = '1.2.0'
   s.license = 'MIT'
   s.summary = 'Swifty API for NSUserDefaults'
   s.homepage = 'https://github.com/radex/SwiftyUserDefaults'
@@ -11,5 +11,5 @@ Pod::Spec.new do |s|
   s.ios.deployment_target = '8.0'
   s.osx.deployment_target = '10.9'
 
-  s.source_files = 'Src/*.swift'
+  s.source_files = 'SwiftyUserDefaults/*.swift'
 end
