+++ SwiftyUserDefaults/SwiftyUserDefaults.h
@@ -1,0 +1,19 @@
+//
+//  SwiftyUserDefaults.h
+//  SwiftyUserDefaults
+//
+//  Created by Alexey Poimtsev on 21/05/15.
+//
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for SwiftyUserDefaults.
+FOUNDATION_EXPORT double SwiftyUserDefaultsVersionNumber;
+
+//! Project version string for SwiftyUserDefaults.
+FOUNDATION_EXPORT const unsigned char SwiftyUserDefaultsVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <SwiftyUserDefaults/PublicHeader.h>
+
+
